mc rate controller: I term gain
Reduce the I gain for high rate errors to reduce bounce-back effects after
flips. Up to 200 degrees the gain is almost not reduced (<25%), so this
should only take noticeable effects for large errors (setpoint changes),
where we actually want to have an effect.

This allows to increase the MC_*RATE_I parameters w/o negative effects
when doing flips (i.e. bounce-back after flips).

The 400 degrees limit and the x^2 are empirical values.

The better the rate tracking in general (high P gain), the less this is
required (because of the lower tracking error). But it also does not harm
in that case, as the i_gain will always be close to 1.